Extract netstack/android related functions into individual file
